requirements.txt
asgiref==3.11.1
black==26.1.0
cfgv==3.5.0
click==8.3.1
distlib==0.4.0
django==6.0.3
djangorestframework==3.16.1
filelock==3.25.0
identify==2.6.17
mypy-extensions==1.1.0
nodeenv==1.10.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.2
pre-commit==4.5.1
python-discovery==1.1.0
pytokens==0.4.1
pyyaml==6.0.3
ruff==0.15.4
sqlparse==0.5.5
virtualenv==21.1.0
django-environ==0.13.0
pillow==12.1.1
